moved functions into epoch tracker
